pyproject.toml
[project]
name = "xclaw"
version = "0.4.0"
description = "Vision-based screen parser & controller for LLM agents"
requires-python = ">=3.12"

dependencies = [
    # ── 感知层（跨平台） ──
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "transformers>=4.40.0,<4.46.0",
    "safetensors>=0.4.0",
    "accelerate>=0.30.0",
    "einops>=0.8.0",
    "timm>=1.0.0",
    "supervision>=0.25.0",
    "onnxruntime>=1.17.0",
    "paddleocr>=2.8.0",

    # ── 截屏 + 图像（跨平台） ──
    "mss>=9.0.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",

    # ── 工具（跨平台） ──
    "pyperclip>=1.8.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "huggingface-hub[cli]>=0.25.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
win = [
    "onnxruntime-gpu>=1.17.0",
    "paddlepaddle-gpu>=2.6.0",
    "pywin32>=306",
]
mac = [
    "pyobjc-framework-Quartz>=10.0",
    "paddlepaddle>=3.0.0",
]
dev = [
    "pytest>=9.0.2",
    "pre-commit>=3.0.0",
]

[project.scripts]
xclaw = "xclaw.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

# torch/torchvision: macOS 走默认 PyPI (MPS), Win/Linux 走 cu121
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }]
torchvision = [{ index = "pytorch-cu121", marker = "sys_platform != 'darwin'" }]

[tool.pytest.ini_options]
markers = [
    "gpu: requires GPU and OmniParser models",
    "slow: tests that take more than 5 seconds",
    "integration: full scheduler integration tests",
    "bench: performance benchmarks",
]
addopts = "-m 'not gpu and not bench'"
